pyproject.toml
[project]
name = "pokerbench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorama>=0.4.6",
    "fastapi>=0.115.12",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "texasholdem>=0.11.0",
    "uvicorn>=0.34.3",
]
